model done, api setup

+++ server/models/booking.js
@@ -3,7 +3,7 @@ module.exports = function(sequelize, DataTypes) {
     active: DataTypes.BOOLEAN,
     day: DataTypes.ENUM('Monday', 'Tuesday', 'Wednesday', 'Thursday', 'Friday', 'Saturday', 'Sunday'),
     time: DataTypes.INTEGER,
-    hours: DataTypes.INTEGER,
+    hours: DataTypes.DECIMAL,
     notes: DataTypes.TEXT
   }, {
     classMethods: {
